backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
openai
drf-spectacular
drf-spectacular-sidecar
PyPDF2
